requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
torch>=1.9.0
transformers>=4.11.0
numpy>=1.20.0
tqdm>=4.62.0
Werkzeug>=2.0.0
python-docx>=0.8.11
PyPDF2>=2.0.0